Reduced compiler warnings
Now avoiding warnings even at compiler warning level "More".

+++ src/MI/ModuleInterface.h
@@ -433,7 +433,7 @@ public:
   // Whether this module is active or has not been reachable for a while
   bool is_active() const {
     #ifdef IS_MASTER
-    return comm_failures < MI_INACTIVE_THRESHOLD && get_last_alive_age() != -1 && get_last_alive_age() < 1000ul*MI_INACTIVE_TIME_THRESHOLD;
+    return comm_failures < (uint8_t)MI_INACTIVE_THRESHOLD && get_last_alive_age() != -1 && get_last_alive_age() < 1000L*MI_INACTIVE_TIME_THRESHOLD;
     #else
     return true;
     #endif
